pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-network-launcher"
version = "0.1.0"
description = "Desktop launcher LAN pair-à-pair avec Flet"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "flet==0.82.2",
  "Pillow>=10,<12",
  "pystray>=0.19,<0.20",
]

[project.optional-dependencies]
build = [
  "pyinstaller>=6,<7",
]

[project.scripts]
py-network-launcher = "desktop_entry:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = [
  "app_views",
  "desktop_entry",
  "discovery",
  "executor",
  "launcher_app",
  "launcher_cli",
  "main_window",
  "models",
  "runtime_paths",
  "service",
  "system_tray",
  "startup",
  "storage",
  "ui_components",
]
